osprey/server/requirements.txt
Flask-SQLAlchemy
Flask-Migrate
psycopg2
flask
requests
pandas
proxystore
dill
globus_compute_sdk
globus_sdk